requirements.txt
pandas
matplotlib
pandasql
urwid
pillow
